Castor3D: Added customisation of background render pass.
