+++ include/Core/Castor3D/Scene/Background/Image.hpp
@@ -59,6 +59,12 @@ namespace castor3d
 		*\copydoc	castor3d::SceneBackground::accept
 		*/
 		C3D_API void accept( BackgroundVisitor & visitor )override;
+		/**
+		*\copydoc	castor3d::SceneBackground::write
+		*/
+		C3D_API bool write( castor::String const & tabs
+			, castor::Path const & folder
+			, castor::StringStream & stream )const override;
 
 	private:
 		bool doInitialise( RenderDevice const & device )override;
