+++ source/Core/Castor3D/Render/EnvironmentMap/EnvironmentMapPass.cpp
@@ -9,7 +9,7 @@
 #include "Castor3D/Render/Viewport.hpp"
 #include "Castor3D/Render/Culling/FrustumCuller.hpp"
 #include "Castor3D/Render/EnvironmentMap/EnvironmentMap.hpp"
-#include "Castor3D/Render/Passes/BackgroundPass.hpp"
+#include "Castor3D/Render/Passes/BackgroundRenderer.hpp"
 #include "Castor3D/Render/Technique/ForwardRenderTechniquePass.hpp"
 #include "Castor3D/Render/Ssao/SsaoConfig.hpp"
 #include "Castor3D/Scene/Camera.hpp"
