+++ source/Core/Castor3D/Scene/Background/Skybox.cpp
@@ -22,6 +22,102 @@
 #include <ashespp/Shader/ShaderModule.hpp>
 #include <ashespp/Sync/Fence.hpp>
 
+#include <CastorUtils/Data/TextWriter.hpp>
+
+namespace castor
+{
+	using namespace castor3d;
+
+	template<>
+	class TextWriter< castor3d::SkyboxBackground >
+		: public TextWriterT< castor3d::SkyboxBackground >
+	{
+	public:
+		explicit TextWriter( castor::String const & tabs
+			, Path const & folder )
+			: TextWriterT< SkyboxBackground >{ tabs }
+			, m_folder{ folder }
+		{
+		}
+
+		bool operator()( SkyboxBackground const & background
+			, castor::StringStream & file )override
+		{
+			log::info << tabs() << cuT( "Writing SkyboxBackground" ) << std::endl;
+			static String const faces[]
+			{
+				cuT( "right" ),
+				cuT( "left" ),
+				cuT( "bottom" ),
+				cuT( "top" ),
+				cuT( "back" ),
+				cuT( "front" ),
+			};
+
+			auto result = true;
+			file << ( cuT( "\n" ) + tabs() + cuT( "//Skybox\n" ) );
+
+			if ( !background.getEquiTexturePath().empty()
+				&& castor::File::fileExists( background.getEquiTexturePath() ) )
+			{
+				if ( auto block{ beginBlock( file, "skybox" ) } )
+				{
+					Path subfolder{ cuT( "Textures" ) };
+					String relative = copyFile( background.getEquiTexturePath()
+						, m_folder
+						, subfolder );
+					string::replace( relative, cuT( "\\" ), cuT( "/" ) );
+					auto & size = background.getEquiSize();
+					file << ( tabs() + cuT( "equirectangular" )
+						+ cuT( " \"" ) + relative + cuT( "\" " )
+						+ string::toString( size.getWidth() ) + cuT( "\n" ) );
+					castor::TextWriter< SkyboxBackground >::checkError( result, "Skybox equi-texture" );
+				}
+			}
+			else if ( !background.getCrossTexturePath().empty()
+				&& castor::File::fileExists( background.getCrossTexturePath() ) )
+			{
+				result = false;
+
+				if ( auto block{ beginBlock( file, "skybox" ) } )
+				{
+					result = writeFile( file, cuT( "cross" ), background.getCrossTexturePath(), m_folder, cuT( "Textures" ) );
+				}
+			}
+			else if ( castor::File::fileExists( Path{ background.getTexture().getLayerCubeFaceView( 0, CubeMapFace( 0u ) ).toString() } )
+				&& castor::File::fileExists( Path{ background.getTexture().getLayerCubeFaceView( 0, CubeMapFace( 1u ) ).toString() } )
+				&& castor::File::fileExists( Path{ background.getTexture().getLayerCubeFaceView( 0, CubeMapFace( 2u ) ).toString() } )
+				&& castor::File::fileExists( Path{ background.getTexture().getLayerCubeFaceView( 0, CubeMapFace( 3u ) ).toString() } )
+				&& castor::File::fileExists( Path{ background.getTexture().getLayerCubeFaceView( 0, CubeMapFace( 4u ) ).toString() } )
+				&& castor::File::fileExists( Path{ background.getTexture().getLayerCubeFaceView( 0, CubeMapFace( 5u ) ).toString() } ) )
+			{
+				result = false;
+
+				if ( auto block{ beginBlock( file, "skybox" ) } )
+				{
+					result = true;
+
+					Path subfolder{ cuT( "Textures" ) };
+
+					for ( uint32_t i = 0; i < 6 && result; ++i )
+					{
+						result = writeFile( file
+							, faces[i]
+							, Path{ background.getTexture().getLayerCubeFaceView( 0u, CubeMapFace( i ) ).toString() }
+							, m_folder
+							, cuT( "Textures" ) );
+					}
+				}
+			}
+
+			return result;
+		}
+
+	private:
+		Path m_folder;
+	};
+}
+
 namespace castor3d
 {
 	//************************************************************************************************
@@ -54,7 +150,7 @@ namespace castor3d
 	SkyboxBackground::SkyboxBackground( Engine & engine
 		, Scene & scene
 		, castor::String const & name )
-		: SceneBackground{ engine, scene, name + cuT( "Skybox" ), BackgroundType::eSkybox }
+		: SceneBackground{ engine, scene, name + cuT( "Skybox" ), cuT( "skybox" ) }
 	{
 		m_texture = std::make_shared< TextureLayout >( *engine.getRenderSystem()
 			, skybox::doGetImageCreate( VK_FORMAT_R8G8B8A8_UNORM, { 16u, 16u }, false )
@@ -66,6 +162,13 @@ namespace castor3d
 	void SkyboxBackground::accept( BackgroundVisitor & visitor )
 	{
 		visitor.visit( *this );
+	}
+
+	bool SkyboxBackground::write( castor::String const & tabs
+		, castor::Path const & folder
+		, castor::StringStream & stream )const
+	{
+		return castor::TextWriter< SkyboxBackground >{ tabs, folder }( *this, stream );
 	}
 
 	void SkyboxBackground::loadLeftImage( castor::Path const & folder
