+++ source/Core/SceneExporter/Text/TextBackground.cpp
@@ -1,9 +1,5 @@
 #include "TextBackground.hpp"
 
-#include "TextColourBackground.hpp"
-#include "TextImageBackground.hpp"
-#include "TextSkyboxBackground.hpp"
-
 namespace castor
 {
 	using namespace castor3d;
@@ -18,16 +14,8 @@ namespace castor
 	bool TextWriter< SceneBackground >::operator()( SceneBackground const & background
 		, StringStream & file )
 	{
-		switch ( background.getType() )
-		{
-		case BackgroundType::eColour:
-			return true;
-		case BackgroundType::eImage:
-			return writeSub( file, static_cast< ImageBackground const & >( background ), m_folder );
-		case BackgroundType::eSkybox:
-			return writeSub( file, static_cast< SkyboxBackground const & >( background ), m_folder );
-		default:
-			return false;
-		}
+		auto result = background.write( tabs(), m_folder, file );
+		checkError( result, cuT( "" ) );
+		return result;
 	}
 }
